+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## master
 
 * Added TypeScript definitions
+* When type is wrong, functions throws `TypeError` instead of `Error`
 
 ## 0.1.0